refactor(data): adoptar DTOs de JOOQ y actualizar lógica de negocio
- Se eliminó JPA por completo (anotaciones, dependencias, configuración).
- Interfaces de servicio (`PlantService`, `EnvironmentalMeasurementService`) ahora usan DTOs JOOQ.
- Se actualizó implementación de servicios a los nuevos records y función SQL.
- Se eliminó configuración JPA en `application.yaml`.
