+++ data/src/main/java/com/cyver/plant/data/service/PlantService.java
@@ -2,10 +2,10 @@ package com.cyver.plant.data.service;
 
 import java.util.List;
 
-import com.cyver.plant.commons.dto.PlantDto;
+import com.cyver.plant.database.domain.udt.dtos.PlantWithLastEnvironmentalMeasurement;
 
 public interface PlantService {
 
-    List<PlantDto> getPlantsByOwner(String auth0Sid);
+    List<PlantWithLastEnvironmentalMeasurement> getPlantsByOwner(String auth0Sid);
 
 }
