+++ data/src/main/java/com/cyver/plant/data/service/impl/EnvironmentalMeasurementImpl.java
@@ -6,10 +6,9 @@ import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cyver.plant.commons.dto.EnvironmentalMeasurementDto;
 import com.cyver.plant.data.service.EnvironmentalMeasurementService;
-import com.cyver.plant.database.EnvironmentalMeasurementRepository;
-import com.cyver.plant.utilities.stream.StreamUtilComponent;
+import com.cyver.plant.database.domain.tables.dtos.EnvironmentalMeasurement;
+import com.cyver.plant.database.domain.tables.repositories.EnvironmentalMeasurementRepository;
 
 import lombok.AllArgsConstructor;
 
@@ -19,10 +18,9 @@ public class EnvironmentalMeasurementImpl implements EnvironmentalMeasurementSer
 
     private final EnvironmentalMeasurementRepository environmentalMeasurementRepository;
 
-    private final StreamUtilComponent streamUtilComponent;
-
     @Override
-    public List<EnvironmentalMeasurementDto> getEnvironmentalMeasurementsByPlant(final UUID plantUuId) {
-        return streamUtilComponent.environmentalStreamToListDto(environmentalMeasurementRepository.findAllByPlantUuid(plantUuId).stream());
+    public List<EnvironmentalMeasurement> getEnvironmentalMeasurementsByPlant(final UUID plantId) {
+        return environmentalMeasurementRepository.fetchByPlantUuid(plantId);
     }
+
 }
