+++ data/src/main/java/com/cyver/plant/data/service/impl/PlantServiceImpl.java
@@ -1,17 +1,17 @@
 package com.cyver.plant.data.service.impl;
 
 import java.util.List;
-import java.util.UUID;
 
+import org.jooq.DSLContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cyver.plant.commons.dto.PlantDto;
-import com.cyver.plant.commons.entities.Owner;
 import com.cyver.plant.data.service.PlantService;
-import com.cyver.plant.database.OwnerRepository;
-import com.cyver.plant.database.PlantRepository;
-import com.cyver.plant.utilities.stream.StreamUtilComponent;
+import com.cyver.plant.database.domain.Routines;
+import com.cyver.plant.database.domain.tables.repositories.OwnerRepository;
+import com.cyver.plant.database.domain.tables.dtos.Owner;
+import com.cyver.plant.database.domain.udt.dtos.PlantWithLastEnvironmentalMeasurement;
+import com.cyver.plant.utilities.owner.OwnerUtilComponent;
 
 import lombok.AllArgsConstructor;
 
@@ -19,13 +19,17 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor(onConstructor_ = { @Autowired })
 public class PlantServiceImpl implements PlantService {
 
-    private final PlantRepository plantRepository;
-
     private final OwnerRepository ownerRepository;
 
+    private final OwnerUtilComponent ownerUtilComponent;
+
+    private final DSLContext dslContext;
+
     @Override
-    public List<PlantDto> getPlantsByOwner(final String auth0Sid) {
-        Owner owner = ownerRepository.findByAuth0Sid(auth0Sid).orElseThrow();
-        return plantRepository.findPlantsByOwner(owner.getUuid());
+    public List<PlantWithLastEnvironmentalMeasurement> getPlantsByOwner(final String auth0Sid) {
+        Owner owner = ownerUtilComponent.getOwnerByEmailIfExists(ownerRepository, auth0Sid);
+        return Routines.getPlantWithLastEnvironmentalMeasurement(dslContext.configuration(), owner.ownerUuid())
+                .into(PlantWithLastEnvironmentalMeasurement.class);
     }
+
 }
