+++ data/src/main/java/com/cyver/plant/data/web/rest/PlantResource.java
@@ -1,6 +1,5 @@
 package com.cyver.plant.data.web.rest;
 
-import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,10 +14,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.cyver.plant.commons.dto.EnvironmentalMeasurementDto;
-import com.cyver.plant.commons.dto.PlantDto;
 import com.cyver.plant.data.service.EnvironmentalMeasurementService;
 import com.cyver.plant.data.service.PlantService;
+import com.cyver.plant.database.domain.tables.dtos.EnvironmentalMeasurement;
+import com.cyver.plant.database.domain.udt.dtos.PlantWithLastEnvironmentalMeasurement;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -34,13 +33,13 @@ public class PlantResource {
     private final EnvironmentalMeasurementService environmentalMeasurementService;
 
     @GetMapping(produces = { MediaType.APPLICATION_JSON_VALUE })
-    public ResponseEntity<List<PlantDto>> getPlantsByOwner(@AuthenticationPrincipal OAuth2User principal) {
+    public ResponseEntity<List<PlantWithLastEnvironmentalMeasurement>> getPlantsByOwner(@AuthenticationPrincipal OAuth2User principal) {
         log.info("REST request to get plants by owner : {}", principal.getName());
-        return new ResponseEntity<>(plantService.getPlantsByOwner(principal.getAttribute("sid")), HttpStatus.OK);
+        return new ResponseEntity<>(plantService.getPlantsByOwner(principal.getAttribute("email")), HttpStatus.OK);
     }
 
     @GetMapping(path = "/{plantUuid}/environmental-measurement", produces = { MediaType.APPLICATION_JSON_VALUE })
-    public ResponseEntity<List<EnvironmentalMeasurementDto>> getMeasurementsByPlant(@PathVariable("plantUuid") final UUID plantUuid) {
+    public ResponseEntity<List<EnvironmentalMeasurement>> getMeasurementsByPlant(@PathVariable("plantUuid") final UUID plantUuid) {
         return ResponseEntity.ok(environmentalMeasurementService.getEnvironmentalMeasurementsByPlant(plantUuid));
     }
