+++ data/src/main/resources/application.yaml
@@ -5,12 +5,6 @@ spring:
     username: ${DATABASE_USER}
     password: ${DATABASE_PASSWORD}
     driver-class-name: org.postgresql.Driver
-  jpa:
-    hibernate.ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
 server:
   port: 8081
 api:
